회원제 게시판, 본인 외 수정 삭제 불가

+++ src/main/java/com/example/todaysmenu/board/controller/BoardController.java
@@ -5,6 +5,9 @@ import com.example.todaysmenu.board.entity.Criteria;
 import com.example.todaysmenu.board.entity.PageDTO;
 import com.example.todaysmenu.board.repository.BoardRepository;
 import com.example.todaysmenu.board.service.BoardService;
+import com.example.todaysmenu.member.entity.MemberDTO;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -14,12 +17,19 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+
 
 @Log4j2
 @Controller
 @RequestMapping("/board/*")
 public class BoardController {
+
+    String danger = "btn btn-danger";
+    String success = "btn btn-success";
 
     @Autowired
     public BoardService boardService;
@@ -38,9 +48,34 @@ public class BoardController {
     @GetMapping("/write.do")
     public String write(
             @RequestParam(value = "tfb_seq",required = false)
-            Integer tfb_seq,Model model) {
+            Integer tfb_seq, Model model,RedirectAttributes rttr, HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        MemberDTO memberSession = (MemberDTO) session.getAttribute("memberDTO");
+        BoardDTO boardInfo;
+        String memberWriter;
+        String boardWriter;
+        try{
+            boardInfo = boardService.info(tfb_seq);
+            memberWriter = memberSession.getTmt_memb_name();
+            boardWriter = boardInfo.getTfb_input_nm();
+        } catch (NullPointerException e) {
+            if(memberSession == null) {
+                return redirect(rttr,"실패 메세지","로그인을 해주시길 바랍니다.",danger);
+            }else{
+                return "board/write";
+            }
+        }
+        if(memberWriter.equals(boardWriter) ) {
+            log.info("test");
+        }else{
+            return redirect(rttr,"실패 메세지","본인글만 수정 삭제 가능합니다.",danger);
+        }
+        if(memberSession == null) {
+            return redirect(rttr,"실패 메세지","로그인을 해주시길 바랍니다.",danger);
+        }
         if(tfb_seq != null){
             model.addAttribute("info",boardService.info(tfb_seq));
+            model.addAttribute("memberSession",memberSession);
             return "board/update";
         }else{
             return "board/write";
@@ -48,14 +83,25 @@ public class BoardController {
     }
 
     @PostMapping("/proc.do")
-    public String proc(BoardDTO boardDTO, @ModelAttribute("cri") Criteria cri, RedirectAttributes rttr) {
-        log.info("boardDTO : {}", boardDTO);
+    public String proc(BoardDTO boardDTO, @ModelAttribute("cri") Criteria cri, RedirectAttributes rttr,HttpServletRequest request) {
+
         String  tfb_seq = boardDTO.getTfb_seq();
-        log.info("=======dataSeq========{}",tfb_seq);
+        HttpSession session = request.getSession();
+        MemberDTO memberSession = (MemberDTO) session.getAttribute("memberDTO");
+        boardDTO.setTfb_input_nm(memberSession.getTmt_memb_name());
+        boardDTO.setTfb_moder_nm(memberSession.getTmt_memb_name());
         if(tfb_seq == null){
             boardService.insert(boardDTO);
         }else{
-            boardService.update(boardDTO);
+            int dataSeq = Integer.parseInt(boardDTO.getTfb_seq());
+            String memberWriter = memberSession.getTmt_memb_name();
+            BoardDTO boardInfo = boardService.info(dataSeq);
+            String boardWriter = boardInfo.getTfb_input_nm();
+            if(memberWriter.equals(boardWriter)){
+                boardService.update(boardDTO);
+            }else{
+                return redirect(rttr,"실패 메세지","본인글만 수정 삭제 가능합니다.",danger);
+            }
             rttr.addFlashAttribute("result","success");
             rttr.addAttribute("pageNum",cri.getPageNum());
             rttr.addAttribute("amount",cri.getAmount());
@@ -71,10 +117,30 @@ public class BoardController {
     }
 
     @GetMapping("/delete.do")
-    public String  delete(BoardDTO boardDTO, @ModelAttribute("cri") Criteria cri, RedirectAttributes rttr){
+    public String  delete(BoardDTO boardDTO, @ModelAttribute("cri") Criteria cri, RedirectAttributes rttr, HttpServletRequest request){
         log.info("=======boardDTO========{}",boardDTO);
         int tfb_seq = Integer.parseInt(boardDTO.getTfb_seq());
-        boardService.delete(tfb_seq);
+        HttpSession session = request.getSession();
+        MemberDTO memberSession = (MemberDTO) session.getAttribute("memberDTO");
+
+        BoardDTO boardInfo;
+        String memberWriter;
+        String boardWriter;
+        try{
+            boardInfo = boardService.info(tfb_seq);
+            memberWriter = memberSession.getTmt_memb_name();
+            boardWriter = boardInfo.getTfb_input_nm();
+        } catch (NullPointerException e) {
+            return redirect(rttr,"실패 메세지","로그인을 해주시길 바랍니다.",danger);
+        }
+
+
+        if(memberWriter.equals(boardWriter)){
+            boardService.delete(tfb_seq);
+        }else{
+            return redirect(rttr,"실패 메세지","본인글만 수정 삭제 가능합니다.",danger);
+        }
+
         rttr.addFlashAttribute("result","success");
         rttr.addAttribute("pageNum",cri.getPageNum());
         rttr.addAttribute("amount",cri.getAmount());
@@ -90,6 +156,14 @@ public class BoardController {
         rttr.addFlashAttribute("result","success");
         rttr.addAttribute("pageNum",cri.getPageNum());
         rttr.addAttribute("amount",cri.getAmount());
+        return "redirect:/board/index.do";
+    }
+
+
+    public String redirect(RedirectAttributes rttr, String p1, String p2, String p3) {
+        rttr.addFlashAttribute("msgType",p1);
+        rttr.addFlashAttribute("msg",p2);
+        rttr.addFlashAttribute("result",p3);
         return "redirect:/board/index.do";
     }
 }
